infer: init at 0.15.0
Status:
  Java support is working, requires javac being in PATH.
  C and C++ support are broken, need to setup the clang wrapper.
  Objective-C support is unknown, aka untested.

Currently infer is a binary package. Infer requires compiling a custom version
of clang, as well as some OCaml dependencies are not packaged in nixpkgs. The
goal is to eventually build infer from source once I figure out how to leverage
the current existing infrastructure related to OCaml and LLVM.